# docs/swagger.yaml
basePath: /api
host: localhost:8080
info:
  contact:
    email: z4har.sapan@gmail.com
    name: S4pan
  description: Здесь можно делиться своими мыслями и искать себе кумиров.
  title: Zadan1e
  version: "1.0"
paths:
...
  /api/auth/createORsave{uuid}:
    post:
      description: 'Эндпоинт может быть выглядить: ../save{uuid} или ../create{uuid}'
      parameters:
      - description: Post uuid
        in: path
        name: uuid
        type: string
      - description: Draft or Published
        in: body
        name: status
        required: true
        schema:
          type: string
      responses: {}
      summary: Публикация, изменение или сохранение в черновик поста
      tags:
      - Posts
...
  /api/posts/{uuid}:
    get:
      description: Страница редактирования конкретного поста.
      parameters:
      - description: Post uuid
        in: path
        name: uuid
        type: string
      responses: {}
      summary: Редактирование поста
      tags:
      - Posts
...
swagger: "2.0"